Transaction double fix Servershop is now ready Default config gets rolled out if Servershop is enabled for the first time

+++ src/main/java/com/geNAZt/RegionShop/Interface/Shop/ShopAdd.java
@@ -110,7 +110,7 @@ public class ShopAdd extends ShopCommand {
                         itemName = "(" + itemInHand.getItemMeta().getDisplayName() + ")";
                     }
 
-                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.ADD, region, player.getName(), itemInHand.getTypeId(), itemInHand.getAmount(), sell, buy);
+                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.ADD, region, player.getName(), itemInHand.getTypeId(), itemInHand.getAmount(), sell.doubleValue(), buy.doubleValue());
 
                     player.sendMessage(Chat.getPrefix() + ChatColor.GOLD + "Added "+ ChatColor.GREEN + ItemName.nicer(itemName) + ChatColor.GOLD + " to the shop.");
                     return;
