+++ src/main/java/com/geNAZt/RegionShop/Listener/PlayerDropItem.java
@@ -58,7 +58,7 @@ public class PlayerDropItem implements Listener {
 
                 plugin.getDatabase().update(item);
 
-                Transaction.generateTransaction(e.getPlayer(), ShopTransaction.TransactionType.EQUIP, region, item.getOwner(), item.getItemID(), droppedItem.getAmount(), item.getSell(), item.getBuy());
+                Transaction.generateTransaction(e.getPlayer(), ShopTransaction.TransactionType.EQUIP, region, item.getOwner(), item.getItemID(), droppedItem.getAmount(), item.getSell().doubleValue(), item.getBuy().doubleValue());
 
                 e.getItemDrop().remove();
             } else {
@@ -72,7 +72,7 @@ public class PlayerDropItem implements Listener {
                     itemName = ItemName.getDataName(droppedItem) + droppedItem.getType().toString();
                 }
 
-                Transaction.generateTransaction(e.getPlayer(), ShopTransaction.TransactionType.EQUIP, region, e.getPlayer().getName(), item.getItemID(), droppedItem.getAmount(), 0, 0);
+                Transaction.generateTransaction(e.getPlayer(), ShopTransaction.TransactionType.EQUIP, region, e.getPlayer().getName(), item.getItemID(), droppedItem.getAmount(), 0.0, 0.0);
 
                 e.getPlayer().sendMessage(Chat.getPrefix() + ChatColor.GOLD + "Added "+ ChatColor.GREEN + ItemName.nicer(itemName) + ChatColor.GOLD + " to the shop.");
             }
