+++ src/main/java/com/geNAZt/RegionShop/Storages/SignEquipStorage.java
@@ -80,7 +80,7 @@ public class SignEquipStorage {
                                     item.setCurrentAmount(item.getCurrentAmount() + iStack.getAmount());
                                     plugin.getDatabase().update(item);
 
-                                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.EQUIP, region, owner, iStack.getTypeId(), iStack.getAmount(), item.getSell(), item.getBuy());
+                                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.EQUIP, region, owner, iStack.getTypeId(), iStack.getAmount(), item.getSell().doubleValue(), item.getBuy().doubleValue());
 
                                 } else {
                                     ItemConverter.toDBItem(iStack, plugin.getServer().getWorld(world), owner, region, 0, 0, 0);
@@ -93,7 +93,7 @@ public class SignEquipStorage {
                                         EssentialBridge.sendMail(Chat.getPrefix(), player, ChatColor.GOLD + "A Equip Sign has added a new Item to your Shop");
                                     }
 
-                                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.EQUIP, region, owner, iStack.getTypeId(), iStack.getAmount(), 0, 0);
+                                    Transaction.generateTransaction(player, ShopTransaction.TransactionType.EQUIP, region, owner, iStack.getTypeId(), iStack.getAmount(), 0.0, 0.0);
 
                                 }
